Added SendRaw with byte data. Closes #288

+++ examples/IRrecord/IRrecord.ino
@@ -65,7 +65,7 @@ void setup() {
 int codeType = -1; // The type of code
 uint32_t codeValue; // The code value if not raw
 uint16_t address; // The address value if not raw
-unsigned int rawCodes[RAW_BUFFER_LENGTH]; // The durations if raw
+uint16_t rawCodes[RAW_BUFFER_LENGTH]; // The durations if raw
 uint8_t codeLen; // The length of the code
 int toggle = 0; // The RC5/6 toggle state
